Moved experiment helper functions into separate file
